Ui class: extract Ui behaviour
Ui display functions is handled by Aurora.java.

Extracting these functions could improve code readability and reduce
coupling.

Let's
* Create a new class Ui
* Extract print msg
* Extract print taskList
